Safe unlink to avoid warnings

+++ tests/ApplicationTest.php
@@ -30,8 +30,13 @@ class ApplicationTest extends TestCase
 
     public function tearDown(): void
     {
-        unlink(static::$in);
-        unlink(static::$ou);
+        // Make sure we clean up after ourselves:
+        if (file_exists(static::$in)) {
+            unlink(static::$in);
+        }
+        if (file_exists(static::$ou)) {
+            unlink(static::$ou);
+        }
     }
 
     public function test_new()
