+++ tests/Helper/OutputHelperTest.php
@@ -34,7 +34,10 @@ class OutputHelperTest extends TestCase
 
     public static function tearDownAfterClass(): void
     {
-        unlink(static::$ou);
+        // Make sure we clean up after ourselves:
+        if (file_exists(static::$ou)) {
+            unlink(static::$ou);
+        }
     }
 
     public function test_show_arguments()
